// src/interfaces/user.ts
export interface Iuser {
    Employee_ID_Number: number;
    Personal_ID_Number: number;
    First_Name: string;
    Middle_Name: string;
    Last_Name: string;
    Day_of_Birth: number;
    Month_of_Birth: number;
    Year_of_Birth: number
    Cellular_Phone: number;
    Home_Phone: number;
    City: string;
    Address: string;
    Postal_Code: number;
    Qualification: string;
    Current_Experience: string;
    Start_Date_Day: number;
    Start_Date_Month: number;
    Start_Date_Year: number;
    End_Date_Day: number;
    End_Date_Month: number;
    End_Date_Year: number;
    Type_of_Employee: string;
    Gender: string;
    Marital_Status: string;
}